fix: Ensure transpiled file is written to disk
Previously, nutra-babel and nutra-coverage did not account for the
possibility of a second transpiler. This resulted in the two
plugins being tighly coupled in the sense that nutra-coverage was
the one writing the file to disk. This commit decouples nutra-babel
by having it write the file to disk and pass on the filename.